Move assets loading in a specific class
